requirements.txt
crewai
boto3
python-dotenv
pydantic
moto
pytest
langchain-google-genai
fastapi
uvicorn
